#153 [update] channel socket update
