+++ frontend/src/components/chat/group/ChannelComponent.tsx
@@ -1,7 +1,8 @@
-import { Box, Grid } from "@mui/material";
-import React, { useState } from "react";
+import { Box, CircularProgress, Grid } from "@mui/material";
+import React, { useEffect, useState } from "react";
 import { Outlet } from "react-router-dom";
 import { Socket } from "socket.io-client";
+import axios from "axios";
 import { ChatRoom } from "../../../types/PrismaType";
 import ChannelGroupComponent from "./ChannelGroupComponent";
 import useSocket from "../../../hooks/useSocket";
@@ -13,6 +14,35 @@ import useSocket from "../../../hooks/useSocket";
 export default function ChannelComponent() {
   const socket: Socket = useSocket('http://localhost:8080/chat');
   const [channels, setChannels] = useState<ChatRoom[]>([]);
+  const [Loading, setLoading] = useState(true);
+
+  const getChannels = async (): Promise<ChatRoom[]> => {
+    const res = await axios.get(`http://localhost:8080/chat/channel`);
+    return res.data;
+  }
+
+  useEffect(() => {
+    try {
+      setLoading(true);
+      getChannels().then((data) => { setChannels(data); })
+    } catch (error) {
+      alert('チャンネル取得に失敗しました。ブラウザをリフレッシュしてください');
+    } finally {
+      setTimeout(() => {
+        setLoading(false);
+      }, 300);
+    }
+  }, [])
+
+  if (Loading) {
+    return (
+      <Box height={'100vh'} sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center' }}>
+        <Box height={'10vh'} width={'10vw'}>
+          <CircularProgress />
+        </Box>
+      </Box>
+    )
+  }
 
   return (
     <>
